requirements.txt
flask
flask-wtf
werkzeug
pycryptodome
flask-limiter
python-dotenv
